Actualizaciones Billon
Actualización de proyecto a versión más reciente 2.2.4

+++ fapp-sii-client-rest/src/main/java/cl/fapp/siiclient/boletas/domain/envio/DteEnvioResponse.java
@@ -17,7 +17,7 @@ public class DteEnvioResponse {
 
     String rut_emisor;
     String rut_envia;
-    Integer trackid;
+    Long trackid;
     String fecha_recepcion;
     String estado;
     String file;
